requirements.txt
pyfiglet
pytest-mock